Refining telemetry and messaging fully working (#24)

+++ back/rbkApiModules/rbkApiModules.Commons.Core/Events/Events/Registry/IEventTypeRegistry.cs
@@ -4,5 +4,5 @@ namespace rbkApiModules.Commons.Core;
 
 public interface IEventTypeRegistry
 {
-    bool TryResolve(string name, int version, out Type type);
+    bool TryResolve(string name, short version, out Type type);
 }
